* commit '11b5befee8a260bba758ef8be4fa6a4975a0fa15':
  Some minor and formating cleanning in snmp probes
  SnmpConnection sometime failed to initialize, try to catch this early
  Minor improvement in Util.parseStringNumber tests.
  Slighlty improved number parsing in xml documents
  Avoid bug in AIX’s snmp agent that returns null value for not found OID.
  Fix IE7 (and IE8 in compatibility mode) time selection mode alignment.
  Fix hours selection limits when selecting the same day, set no limits when selecting different days.
  "dijit/form/TimeTextBox" should not be in global require
  Fix left panel not being full page height.
  Removed useless require on « dijit/form »
  Use explicit require for some dijit modules to make things work on IE8
  Fix malformed html
  Continue with IE javascript errors.
  Fix IE javascript error (object attributes named class needs to be quoted).
  Unit for count and time in JVM GC should be SI and not binary
  Updated Oracle's JVM gc probes.
  Typo in  a ds name in MysqlStatusOperations graph (thanks to drone58)
  Wrong name for a memory pool in JMXSerialGC (thanks to drone58)

+++ src/jrds/starter/XmlProvider.java
@@ -102,11 +102,11 @@ public class XmlProvider extends Starter {
                 if(xpath == null || "".equals(xpath))
                     continue;
                 Node n = (Node)xpather.evaluate(xpath, d, XPathConstants.NODE);
-                double value = 0;
+                Double value = 0.0d;
                 if(n != null) {
                     log(Level.TRACE, "%s", n);
-                    value = jrds.Util.parseStringNumber(n.getTextContent(), Double.NaN).doubleValue();
-                    oldMap.put(xpath, Double.valueOf(value));
+                    value = jrds.Util.parseStringNumber(n.getTextContent(), Double.NaN);
+                    oldMap.put(xpath, value);
                 }
             } catch (XPathExpressionException e) {
                 log(Level.ERROR, "Invalid XPATH : " + xpath + " for " + this);
